Changing sound script to be reusable across all terrain
The sound script now checks for distance between whatever terrain it is attatched to, and the player when refrences are hooked up. Then plays the audiosource attatched to that terrain. It also turns the sound off if they are far away from it to prevent overlapping sounds or odd sounds at inappropriate times.