Fixed position of character so they can be closer to the middle of logs when moving
The movement is grid based so it moves one grid space at a time. I moved the player to be furhter up in the grid space so he landed further up on the next. This put the player in a better spot when crossing the logs. Player can no longer fall off terrain sides.