Changed spawners around for the forest, changed spawn rates, and made some speed changes to cars/logs.
I increased the number of tree spawners by x4, as well as moved them so that they properly blocked the player's movement.  The spawn rate of the trees was also increased from 66% to 80%, to accomodate the larger number of spawners.  The speed of the logs was increased to 3.5 from 3, and the speed of the cars was increased to 5 from 3.